+++ bin/stop-kafka.sh
@@ -3,11 +3,16 @@
 # Script to stop docker
 
 echo "Stopping kafka-docker container"
-# docker stop kafka-docker
-docker kill kafka-docker
+source config.sh
 
-echo "Removing data folder"
-rm -fr data/kafka
+running=$(docker ps | grep " $kafka_docker_container ")
+if [ ! -z "$running" ]; then
+    # docker stop kafka-docker
+    docker kill $kafka_docker_container
 
-echo "Removing kafka-docker container"
-docker rm kafka-docker
+    echo "Removing data folder"
+    rm -fr data/kafka
+
+    echo "Removing kafka-docker container"
+    docker rm $kafka_docker_container
+fi
